fix: Use defined media type for live radio detection, relocate DynamicMediaSourceFactory
